pyproject.toml
[project]
name = "land-rag-toolbox"
version = "0.1.0"
description = "Modular Python RAG framework for LAND-UFRJ"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "chromadb",
    "fastembed",
    "flashrank",
    "ragas",
    "nougat-ocr",
    "python-doctr",
    "python-dotenv",
    "rank_bm25",
    "langchain_text_splitters",
    "pytest",
    "pytest-mock",
    "langgraph",
    "pydantic",
    "transformers",
    "torch"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/land_rag"]
